+++ apps/waitlist/app/api/frame/[fid]/actions/goto-builders/route.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-case-declarations */
 import { deterministicV4UUIDFromFid } from '@connect-shared/lib/farcaster/uuidFromFid';
 import { baseUrl } from '@root/config/constants';
-import { validateFrameInteractionViaAirstackWithErrorCatching } from '@root/lib/farcaster/moxie';
+import { validateFrameInteractionViaAirstackWithErrorCatching } from '@root/lib/farcaster/airstack';
 import type { FarcasterFrameInteractionToValidate } from '@root/lib/farcaster/validateFrameInteraction';
 import { validateFrameInteraction } from '@root/lib/farcaster/validateFrameInteraction';
