+++ apps/waitlist/app/api/frame/[fid]/level-changed/route.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-case-declarations */
 import { InvalidInputError } from '@charmverse/core/errors';
 import { deterministicV4UUIDFromFid } from '@connect-shared/lib/farcaster/uuidFromFid';
-import { validateFrameInteractionViaAirstackWithErrorCatching } from '@root/lib/farcaster/moxie';
+import { validateFrameInteractionViaAirstackWithErrorCatching } from '@root/lib/farcaster/airstack';
 import type { FarcasterFrameInteractionToValidate } from '@root/lib/farcaster/validateFrameInteraction';
 import { validateFrameInteraction } from '@root/lib/farcaster/validateFrameInteraction';
